+++ src/main/java/com/hodor/web/PetController.java
@@ -36,8 +36,8 @@ public class PetController {
      */
     @GetMapping("/user/pets")
     public JsonResult<Map<String, Object>> getUserListByQuery(@RequestParam(required = false) String query,
-                                                              @RequestParam(required = false, defaultValue = "1") Long pagenum,
-                                                              @RequestParam(required = false, defaultValue = "10") Long pagesize) {
+                                                              @RequestParam(required = false, defaultValue = "1") Integer pagenum,
+                                                              @RequestParam(required = false, defaultValue = "10") Integer pagesize) {
         try {
             JsonResult<Map<String, Object>> petListByQuery = petService.getPetListByQuery(query, pagenum, pagesize);
             return petListByQuery;
