+++ src/main/java/com/hodor/web/ReserveController.java
@@ -81,8 +81,8 @@ public class ReserveController {
      */
     @GetMapping("/user/applies")
     public JsonResult<Map<String, Object>> getUserListByQuery(@RequestParam(required = false) String query,
-                                                              @RequestParam(required = false, defaultValue = "1") Long pagenum,
-                                                              @RequestParam(required = false, defaultValue = "10") Long pagesize) {
+                                                              @RequestParam(required = false, defaultValue = "1") Integer pagenum,
+                                                              @RequestParam(required = false, defaultValue = "10") Integer pagesize) {
         try {
             JsonResult<Map<String, Object>> reserveListByQuery = reserveService.getReserveListByQuery(query, pagenum, pagesize);
             return reserveListByQuery;
